[PRMT-4566] Reorganise code, use CreatedAt instead of UpdatedAt to determine latest record
